% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgtime.R
\name{msgtime}
\alias{msgtime}
\title{Overwrite [message()], [warning()], and [stop()] to prepend a time
stamp to their output.}
\usage{
msgtime(enable)
}
\arguments{
\item{enable}{Whether to enable or disable the time
stamping. Setting this to `FALSE` will restore the original
functions.}
}
\description{
This function overwrites the functions in the `base` package rather
than merely masking them, so all packages are affected.
}
\note{
This is potentially dangerous, since it involves messing
  around with the base package's namespace. The functionality of
  warning() and stop() is also slightly degraded, since their
  immediate calling environment is changed to the wrapper function.
}
\examples{
message("Hello.")
warning("Hello?")
\dontrun{
stop("Hello!")
}
msgtime(TRUE)
message("Hello.")
warning("Hello?")
\dontrun{
stop("Hello!")
}
msgtime(FALSE)
message("Hello.")
}
